feat: implement custom Star Nav time slider, show daylight on the track [PT-185463590] [PT-185501241]
