+++ common/package.json
@@ -95,10 +95,9 @@
     "react-dom": "^17.0.2"
   },
   "dependencies": {
-    "@dnd-kit/core": "^5.0.3",
-    "@emotion/react": "^11.9.0",
-    "@emotion/styled": "^11.8.1",
-    "@mui/material": "^5.6.2",
+    "@emotion/react": "^11.11.1",
+    "@emotion/styled": "^11.11.0",
+    "@mui/material": "^5.14.0",
     "@types/deep-equal": "^1.0.1",
     "clsx": "^1.2.1",
     "deep-equal": "^2.0.5",
