--- common/src/components/controls/slider.scss
@@ -1,96 +1,0 @@
-@import "../../vars.scss";
-
-$railHeight: 6px;
-$thumbSize: 40px;
-
-.slider {
-  width: 100%;
-  // Make sure that interacting with slider on touch devices won't trigger scrolling.
-  // https://docs.dndkit.com/api-documentation/sensors/pointer#touch-action
-  touch-action: none;
-
-  &.disabled {
-    opacity: 0.35;
-    pointer-events: none;
-  }
-
-  .label {
-    font-size: 16px;
-    font-weight: bold;
-    color: $bluePrimary;
-    text-align: center;
-    margin-bottom: 6px;
-  }
-
-  :global(.Mui-focusVisible) {
-    outline: $focusOutline;
-  }
-
-  :global(.MuiSlider-root) {
-    height: $railHeight;
-    width: calc(100% - $thumbSize);
-    padding: 17px 0;
-    margin: 0 $thumbSize * 0.5;
-    display: inline-block;
-    position: relative;
-    cursor: pointer;
-
-    :global(.MuiSlider-rail) {
-      display: block;
-      position: absolute;
-      width: 100%;
-      height: $railHeight;
-      border-radius: $railHeight * 0.5;
-      background-color: $purpleTint;
-    }
-
-    :global(.MuiSlider-track) {
-      display: block;
-      position: absolute;
-      height: $railHeight;
-      border-radius: $railHeight * 0.5;
-      background-color: $bluePrimary;
-    }
-
-    :global(.MuiSlider-thumb) {
-      position: absolute;
-      width: $thumbSize;
-      height: $thumbSize;
-      margin-left: -$thumbSize * 0.5;
-      margin-top: -17px;
-      box-sizing: border-box;
-      border-radius: 50%;
-      border: 3px solid $bluePrimary;
-      background-color: #fff;
-
-      &:hover {
-        background-color: $lightBlueTint10;
-      }
-      &:active {
-        background-color: $bluePrimary;
-      }
-    }
-
-    :global(.MuiSlider-mark) {
-      width: 4px;
-      height: 4px;
-      margin-left: -2px;
-      border-radius: 50%;
-      background-color: $bluePrimary;
-      position: absolute;
-      margin-top: 10px;
-    }
-
-    :global(.MuiSlider-markLabel) {
-      font-size: 12px;
-      position: absolute;
-      top: 35px;
-      -webkit-transform: translateX(-50%);
-      -moz-transform: translateX(-50%);
-      -ms-transform: translateX(-50%);
-      transform: translateX(-50%);
-      margin-top: 8px;
-      color: $bluePrimary;
-    }
-  }
-}
