+++ common/src/vars.scss
@@ -52,4 +52,7 @@ $dark-blue: #1e2749;
 $light-blue: #e7edf6;
 $light-blue2: #f7fafe;
 
+$blue-hover: darken($light-blue, 10%);
+$blue-active: darken($light-blue, 20%);
+
 $maxAppWidth: 900px;
