+++ plant-growth/package.json
@@ -107,7 +107,7 @@
     "sass": "^1.49.9",
     "sass-loader": "^12.6.0",
     "script-loader": "^0.7.2",
-    "style-loader": "^3.3.1",
+    "style-loader": "^3.3.3",
     "ts-jest": "^27.1.3",
     "ts-loader": "^9.2.8",
     "ts-node": "^10.7.0",
@@ -122,9 +122,9 @@
   "dependencies": {
     "@concord-consortium/lara-interactive-api": "^1.7.2",
     "@dnd-kit/core": "^5.0.3",
-    "@emotion/react": "^11.9.0",
-    "@emotion/styled": "^11.8.1",
-    "@mui/material": "^5.6.2",
+    "@emotion/react": "^11.11.1",
+    "@emotion/styled": "^11.11.0",
+    "@mui/material": "^5.14.0",
     "@types/deep-equal": "^1.0.1",
     "clsx": "^1.2.1",
     "common": "*",
