+++ star-navigation/package.json
@@ -114,6 +114,9 @@
     "zip-webpack-plugin": "^4.0.1"
   },
   "dependencies": {
+    "@emotion/react": "^11.11.1",
+    "@emotion/styled": "^11.11.0",
+    "@mui/material": "^5.14.0",
     "camera-controls": "^1.36.2",
     "clsx": "^1.2.1",
     "common": "*",
