+++ star-navigation/src/components/bottom-container.scss
@@ -3,11 +3,20 @@
 .bottomContainer {
   display: flex;
   flex-direction: column;
-  height: 160px;
 
   .timeSliderContainer {
     width: 100%;
+    height: 64px;
     background-color: $lightYellow;
+    display: flex;
+    flex-direction: row;
+    align-items: center;
+    justify-content: center;
+
+    .label {
+      font-size: 14px;
+      margin: 0 5px;
+    }
   }
 
   .row {
@@ -38,7 +47,7 @@
 
       .content {
         width: 100%;
-        height: 75px;
+        height: 72px;
         display: flex;
         flex-direction: row;
         align-items: center;
