+++ star-navigation/src/components/vars.scss
@@ -3,6 +3,8 @@
 $darkYellow: #fabc2a;
 $lightYellow: #fff6e1;
 
+$midnight-blue: #0c274c;
+
 // If TS file needs to have an access to a CSS variable, it can be exported:
 // $plot-width: 200px;
 // :export {
